Create file watcher for master nodes
This commit moves the logic of watchign a file only on the master node
into a new shared class, MasterNodeFileWatchingService.

relates #101572
